\name{AEJapp}
\alias{AEJapp}
\docType{data}
\title{
Replication data for: Iron Deficiency and Schooling Attainment in Peru (Chong et al, 2016)
}
\description{
The data is taken from Chong et al. (2016), who study the effect of iron deficiency anemia
  (i.e., anemia caused by a lack of iron) on school-age children’s educational attainment and       cognitive ability in Peru.
}
\usage{data("AEJapp")}
\format{
  A data frame with 215 observations on the 62 variables.
}
\source{
Chong, Alberto, Cohen, Isabelle, Field, Erica, Nakasone, Eduardo, and Torero, Maximo.           Replication data for: Iron Deficiency and Schooling Attainment in Peru. Nashville, TN: American   Economic Association [publisher], 2016. Ann Arbor, MI: Inter-university Consortium for Political  and Social Research [distributor], 2019-10-12. https://doi.org/10.3886/E113624V1.
}
\references{
Chong, Alberto, Isabelle Cohen, Erica Field, Eduardo Nakasone, and Maximo Torero. “Iron     Deficiency and Schooling Attainment in Peru.” American Economic Journal: Applied Economics 8, no.  4 (October 2016): 222–55. https://doi.org/10.1257/app.20140494.
}
\examples{
data(AEJapp)
}
\keyword{datasets}
