% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{print.sreg}
\alias{print.sreg}
\title{Print \code{sreg} Objects}
\usage{
\method{print}{sreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sreg}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
No return value, called for side effects.
}
\description{
Print the summary table of estimation results for \code{sreg} objects.
}
\examples{
set.seed(21)
data <- sreg.rgen(
  n = 80, tau.vec = c(0.2, 0.5), n.strata = 4,
  cluster = FALSE
)
fit <- sreg(
  Y = data$Y, S = data$S, D = data$D,
  X = data[c("x_1", "x_2")]
)
print(fit)
}
