% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sreg-package}
\alias{sreg}
\alias{sreg-package}
\title{sreg: Sudski registar}
\description{
Povlacenje podataka sa sudskog registra.
}
\author{
\strong{Maintainer}: Mislav Sagovac \email{mislav.sagovac@contetnio.biz}

}
