\name{AEJapp}
\alias{AEJapp}
\docType{data}
\title{
Replication data for: Iron Deficiency and Schooling Attainment in Peru (Chong et al, 2016)
}
\description{
The data is taken from Chong et al. (2016), who study the effect of iron deficiency anemia
  (i.e., anemia caused by a lack of iron) on school-age children’s educational attainment and cognitive ability in Peru.
}
\usage{data("AEJapp")}
\format{
  A data frame with 215 observations on the 62 variables.
}
\source{
Chong, A., Cohen, I., Field, E., Nakasone, E., and Torero, M. (2016). Replication data for: Iron Deficiency and Schooling Attainment in Peru. Nashville, TN: American Economic Association [publisher], 2016. Ann Arbor, MI: Inter-university Consortium for Political and Social Research [distributor], 2019-10-12. \doi{10.3886/E113624V1}.
}
\references{
Chong, A., Cohen, I., Field, E., Nakasone, E., and Torero, M. (2016). Iron Deficiency and Schooling Attainment in Peru. \emph{American Economic Journal: Applied Economics}, 8(4), 222–255. \doi{10.1257/app.20140494}.
}
\examples{
data(AEJapp)
}
\keyword{datasets}
