% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{plot.sreg}
\alias{plot.sreg}
\title{Plot Method for `sreg` Objects}
\usage{
\method{plot}{sreg}(x, treatment_labels = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{sreg}.}

\item{treatment_labels}{Optional vector of treatment labels.}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object (invisible).
}
\description{
Visualize estimated ATEs and confidence intervals for objects of class \code{sreg}.
}
