% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.dev.R
\name{sreg}
\alias{sreg}
\title{Estimates the ATE}
\usage{
sreg(Y, S, D, G.id = NULL, Ng = NULL, X = NULL, Ng.cov = FALSE)
}
\arguments{
\item{Y}{a numeric vector of the observed outcomes}

\item{S}{a numeric vector of strata indicators}

\item{D}{a numeric vector of treatments}

\item{G.id}{a numeric vector of cluster indicators}

\item{Ng}{a numeric vector of cluster sizes}

\item{X}{a data frame of covariates}

\item{Ng.cov}{a TRUE/FALSE argument indicating whether the Ng should be included as the only covariate in linear adjustments
when X = NULL}
}
\value{
a list containing the results
}
\description{
Estimates the ATE
}
\examples{
test <- sreg(Y,S,D,X)
}
