% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.dev.R
\name{sreg}
\alias{sreg}
\title{Estimates the ATE}
\usage{
sreg(Y, S = NULL, D, G.id = NULL, Ng = NULL, X = NULL, HC1 = FALSE)
}
\arguments{
\item{Y}{a numeric vector of the observed outcomes}

\item{S}{a numeric vector of strata indicators}

\item{D}{a numeric vector of treatments}

\item{G.id}{a numeric vector of cluster indicators}

\item{Ng}{a numeric vector of cluster sizes}

\item{X}{a data frame of covariates}

\item{HC1}{a TRUE/FALSE argument indicating whether the small sample correction should be applied to the variance estimator.}
}
\value{
a list containing the results
}
\description{
Estimates the ATE
}
\examples{
library("sreg")
library("dplyr")
library("haven")
### Example 1. Simulated Data.
data <- sreg.rgen(n = 1000, tau.vec = c(0), n.strata = 4, cluster = FALSE)
Y <- data$Y
S <- data$S
D <- data$D
X <- data.frame("x_1" = data$x_1, "x_2" = data$x_2)
result <- sreg(Y, S, D, G.id = NULL, Ng = NULL, X)
### Example 2. Data taken from Chong et al. (2016).
## Data description
?AEJapp
## Upload the data from the package
data("AEJapp")
data <- AEJapp
head(data)
## Replicate the empirical illustration from (Bugni et al, 2019)
# Prepare the data
Y <- data$gradesq34
D <- data$treatment
S <- data$class_level
data.clean <- data.frame(Y, D, S)
data.clean <- data.clean \%>\%
  mutate(D = ifelse(D == 3, 0, D))
Y <- data.clean$Y
D <- data.clean$D
S <- data.clean$S
# Look at the frequency table
table(D = data.clean$D, S = data.clean$S)
# Replicate the results from (Bugni et al, 2019)
result <- sreg::sreg(Y, S, D, HC1 = TRUE)

## Besides that, it is possible to add linear adjustments (covariates)
x_1 <- data$pills_taken
x_2 <- data$age_months
data.clean <- data.frame(Y, D, S, x_1, x_2)
data.clean <- data.clean \%>\%
  mutate(D = ifelse(D == 3, 0, D))
Y <- data.clean$Y
D <- data.clean$D
S <- data.clean$S
X <- data.frame(data.clean$x_1, data.clean$x_2)
result <- sreg::sreg(Y, S, D, X = X, HC1 = TRUE)
}
