% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.dev.R
\name{sreg}
\alias{sreg}
\title{Estimate Average Treatment Effects (ATEs) and Corresponding Standard Errors}
\usage{
sreg(Y, S = NULL, D, G.id = NULL, Ng = NULL, X = NULL, HC1 = TRUE)
}
\arguments{
\item{Y}{a numeric \eqn{n \times 1} vector of the observed outcomes}

\item{S}{a numeric \eqn{n \times 1} vector of strata indicators;  if \code{NULL} then the estimation is performed assuming no stratification}

\item{D}{a numeric \eqn{n \times 1} vector of treatments indexed by \eqn{\{0, 1, 2, \ldots\}}, where \eqn{\code{D} = 0} denotes the control}

\item{G.id}{a numeric \eqn{n \times 1} vector of cluster indicators; if \code{NULL} then estimation is performed assuming treatment is assigned at the individual level}

\item{Ng}{a numeric \eqn{n \times 1} vector of cluster sizes; if \code{NULL} then \code{Ng} is assumed to be equal to the number of available observations in every cluster}

\item{X}{a data frame with columns representing the covariate values for every observation; if \code{NULL} then the estimator without linear adjustments is applied. (Note: sreg cannot use individual-level covariates for covariate adjustment in cluster-randomized experiments. Any individual-level covariates will be aggregated to their cluster-level averages)}

\item{HC1}{a \code{TRUE/FALSE} logical argument indicating whether the small sample correction should be applied to the variance estimator}
}
\value{
An object of class \code{sreg} that is a list containing the following elements:
\itemize{
\item \code{tau.hat}: a \eqn{1 \times |\mathcal A|} vector of ATE estimates, where \eqn{|\mathcal A|} represents the number of treatments
\item \code{se.rob}: a \eqn{1 \times |\mathcal A|} vector of standard errors estimates, where \eqn{|\mathcal A|} represents the number of treatments
\item \code{t.stat}: a \eqn{1 \times |\mathcal A|} vector of \eqn{t}-statistics, where \eqn{|\mathcal A|} represents the number of treatments
\item \code{p.value}: a \eqn{1 \times |\mathcal A|} vector of corresponding \eqn{p}-values, where \eqn{|\mathcal A|} represents the number of treatments
\item \code{CI.left}: a \eqn{1 \times |\mathcal A|} vector of the left bounds of the 95\% as. confidence interval
\item \code{CI.right}: a \eqn{1 \times |\mathcal A|} vector of the right bounds of the 95\% as. confidence interval
\item \code{data}: an original data of the form \code{data.frame(Y, S, D, G.id, Ng, X)}
\item \code{lin.adj}: a data frame representing the covariates that were used in implementing linear adjustments
}
}
\description{
This function estimates the ATE(s) and the corresponding standard error(s) for a (collection of) treatment(s) relative to a control.
}
\examples{
library("sreg")
library("dplyr")
library("haven")
### Example 1. Simulated Data.
data <- sreg.rgen(n = 1000, tau.vec = c(0), n.strata = 4, cluster = FALSE)
Y <- data$Y
S <- data$S
D <- data$D
X <- data.frame("x_1" = data$x_1, "x_2" = data$x_2)
result <- sreg(Y, S, D, G.id = NULL, Ng = NULL, X)
### Example 2. Data taken from Chong et al. (2016).
## Data description
?AEJapp
## Upload the data from the package
data("AEJapp")
data <- AEJapp
head(data)
## Replicate the empirical illustration from (Bugni et al, 2019)
# Prepare the data
Y <- data$gradesq34
D <- data$treatment
S <- data$class_level
data.clean <- data.frame(Y, D, S)
data.clean <- data.clean \%>\%
  mutate(D = ifelse(D == 3, 0, D))
Y <- data.clean$Y
D <- data.clean$D
S <- data.clean$S
# Look at the frequency table
table(D = data.clean$D, S = data.clean$S)
# Replicate the results from (Bugni et al, 2019)
result <- sreg::sreg(Y, S, D)

## Besides that, it is possible to add linear adjustments (covariates)
pills <- data$pills_taken
age <- data$age_months
data.clean <- data.frame(Y, D, S, pills, age)
data.clean <- data.clean \%>\%
  mutate(D = ifelse(D == 3, 0, D))
Y <- data.clean$Y
D <- data.clean$D
S <- data.clean$S
X <- data.frame("pills" = data.clean$pills, "age" = data.clean$age)
result <- sreg::sreg(Y, S, D, G.id = NULL, X = X)
}
