% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sreg.R
\name{Sreg}
\alias{Sreg}
\title{Sreg Class}
\description{
Get data from Sudski registar: https://sudreg-data.gov.hr/ords/r/srn_rep/vanjski-srn-rep/home
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token_url}}{Token url}

\item{\code{url}}{Base url for endpoints}

\item{\code{user_name}}{User name}

\item{\code{pass}}{Password}

\item{\code{token}}{token}

\item{\code{docs}}{Documentation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sreg-new}{\code{Sreg$new()}}
\item \href{#method-Sreg-get_token}{\code{Sreg$get_token()}}
\item \href{#method-Sreg-check_token_expiry}{\code{Sreg$check_token_expiry()}}
\item \href{#method-Sreg-get}{\code{Sreg$get()}}
\item \href{#method-Sreg-get_wrap}{\code{Sreg$get_wrap()}}
\item \href{#method-Sreg-get_loop}{\code{Sreg$get_loop()}}
\item \href{#method-Sreg-endpoints}{\code{Sreg$endpoints()}}
\item \href{#method-Sreg-endpoint_meta}{\code{Sreg$endpoint_meta()}}
\item \href{#method-Sreg-endpoint_parameters}{\code{Sreg$endpoint_parameters()}}
\item \href{#method-Sreg-clone}{\code{Sreg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-new}{}}}
\subsection{Method \code{new()}}{
Create a new Sreg object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$new(user_name = Sys.getenv("USER"), pass = Sys.getenv("PASS"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_name}}{User name. Recommended to sava in .Renviron as USER.
Otherwise provide string.}

\item{\code{pass}}{Password. Recommended to save in .Renviron as PASS.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Sreg` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-get_token}{}}}
\subsection{Method \code{get_token()}}{
Get token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String - token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-check_token_expiry"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-check_token_expiry}{}}}
\subsection{Method \code{check_token_expiry()}}{
Get new token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$check_token_expiry()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String - token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-get"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-get}{}}}
\subsection{Method \code{get()}}{
Get data from Sudski registar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$get(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{URL to get data from.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data from Sudski registar.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-get_wrap"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-get_wrap}{}}}
\subsection{Method \code{get_wrap()}}{
Get data from Sudski registar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$get_wrap(tag, q = list(limit = 100, only_active = TRUE), clean = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{Tag to get data from.}

\item{\code{q}}{Query parameters.}

\item{\code{clean}}{Clean data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data from Sudski registar.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-get_loop"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-get_loop}{}}}
\subsection{Method \code{get_loop()}}{
Get data from Sudski registar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$get_loop(tag, q = list(), by = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{Tag to get data from.}

\item{\code{q}}{Query parameters.}

\item{\code{by}}{Number of rows to get.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data from Sudski registar.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-endpoints"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-endpoints}{}}}
\subsection{Method \code{endpoints()}}{
List of all endpoints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$endpoints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of all endpoints.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-endpoint_meta"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-endpoint_meta}{}}}
\subsection{Method \code{endpoint_meta()}}{
Get endpoint meta.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$endpoint_meta(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{Endpoint.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Endpoint meta.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-endpoint_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-endpoint_parameters}{}}}
\subsection{Method \code{endpoint_parameters()}}{
Get endpoint parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$endpoint_parameters(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{Endpoint.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Endpoint parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sreg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sreg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sreg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
