% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.dev.R
\name{sreg.rgen}
\alias{sreg.rgen}
\title{Generates a Pseudo-Random Sample under the Stratified Block Randomization
The function generates the observed outcomes, treatment assignments, strata indicators, cluster indicators, cluster sizes, and covariates for estimating the treatment effect within the context of a stratified block randomization design under the covariate-adaptive randomization (CAR).}
\usage{
sreg.rgen(
  n,
  Nmax = 50,
  n.strata,
  tau.vec = c(0),
  gamma.vec = c(0.4, 0.2, 1),
  cluster = TRUE,
  is.cov = TRUE
)
}
\arguments{
\item{n}{number of observations}

\item{Nmax}{maximum size of clusters}

\item{n.strata}{number of strata}

\item{tau.vec}{a numeric vector of treatment effects}

\item{gamma.vec}{a numeric vector of parameters}

\item{cluster}{a \code{TRUE/FALSE} argument indicating whether the dgp should include clusters or not}

\item{is.cov}{a \code{TRUE/FALSE} argument indicating whether the dgp should include covariates or not}
}
\value{
a data frame containing the generated values of \code{Y}, \code{S}, \code{D}, \code{G.id}, \code{Ng}, and \code{X}
}
\description{
Generates a Pseudo-Random Sample under the Stratified Block Randomization
The function generates the observed outcomes, treatment assignments, strata indicators, cluster indicators, cluster sizes, and covariates for estimating the treatment effect within the context of a stratified block randomization design under the covariate-adaptive randomization (CAR).
}
\examples{
data <- sreg.rgen(n = 1000, tau.vec = c(0), n.strata = 4, cluster = TRUE)
}
