% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.dev.R
\name{sreg.rgen}
\alias{sreg.rgen}
\title{Generates a pseudo-random sample for estimating ATE}
\usage{
sreg.rgen(
  n,
  Nmax = 50,
  n.strata,
  tau.vec = c(0),
  gamma.vec = c(0.4, 0.2, 1),
  cluster = TRUE,
  is.cov = TRUE
)
}
\arguments{
\item{n}{number of observations}

\item{Nmax}{maximum size of clusters}

\item{n.strata}{number of strata}

\item{tau.vec}{a numeric vector of treatment effects}

\item{gamma.vec}{a numeric vector of parameters}

\item{cluster}{a TRUE/FALSE argument indicating whether the dgp should include clusters or not}

\item{is.cov}{a TRUE/FALSE argument indicating whether the dgp should include covariates or not}
}
\value{
a data frame containing the results
}
\description{
Generates a pseudo-random sample for estimating ATE
}
\examples{
data <- sreg.rgen(n = 1000, tau.vec = c(0), n.strata = 4, cluster = TRUE)
}
